Extract unavailable period modal into its own component
