+++ src/app/admin/cases/register/components/CaseForm.tsx
@@ -832,4 +832,3 @@ function EstimateModal({ isOpen, onClose, onCalculate }: EstimateModalProps) {
     </div>
   );
 }
-}
