+++ src/app/admin/dispatch/page.tsx
@@ -6,6 +6,7 @@ import AdminAuthGuard from '@/components/AdminAuthGuard';
 import TruckRegistration from '@/components/TruckRegistration';
 import DispatchCalendar from '@/components/DispatchCalendar';
 import TruckAssignmentModal from './components/TruckAssignmentModal';
+import UnavailablePeriodModal from './components/UnavailablePeriodModal';
 import StatusFilter from '@/components/dispatch/StatusFilter';
 import { TruckManagement } from '@/components/dispatch/TruckManagement';
 import WorkerAssignmentView from './views/WorkerAssignmentView';
@@ -1494,704 +1495,6 @@ const UnifiedCaseManagement = ({
   );
 };
 
-// 車両使用不能期間設定モーダルコンポーネント
-interface UnavailablePeriodModalProps {
-  truck: Truck | null;
-  onClose: () => void;
-  onSave: (period: { startDate: string; endDate: string; reason: string }) => void;
-  initialPeriod: { startDate: string; endDate: string; reason: string };
-  onPeriodChange: (period: { startDate: string; endDate: string; reason: string }) => void;
-}
-
-const UnavailablePeriodModal = ({ truck, onClose, onSave, initialPeriod, onPeriodChange }: UnavailablePeriodModalProps) => {
-  if (!truck) return null;
-  
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    if (!initialPeriod.startDate || !initialPeriod.endDate) {
-      alert('開始日と終了日を入力してください');
-      return;
-    }
-    
-    if (new Date(initialPeriod.startDate) > new Date(initialPeriod.endDate)) {
-      alert('開始日は終了日より前の日付を選択してください');
-      return;
-    }
-    
-    onSave(initialPeriod);
-  };
-  
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md">
-        <h3 className="text-lg font-semibold mb-4">車両使用不能期間設定</h3>
-        <p className="text-sm text-gray-600 mb-4">
-          対象車両: {truck.name} ({truck.plateNumber})
-        </p>
-        
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              開始日 <span className="text-red-500">*</span>
-            </label>
-            <input
-              type="date"
-              value={initialPeriod.startDate}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, startDate: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-              required
-            />
-          </div>
-          
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              終了日 <span className="text-red-500">*</span>
-            </label>
-            <input
-              type="date"
-              value={initialPeriod.endDate}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, endDate: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-              required
-            />
-          </div>
-          
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              理由
-            </label>
-            <select
-              value={initialPeriod.reason}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, reason: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              <option value="">選択してください</option>
-              <option value="車検">車検</option>
-              <option value="定期点検">定期点検</option>
-              <option value="修理">修理</option>
-              <option value="清掃・メンテナンス">清掃・メンテナンス</option>
-              <option value="休車">休車</option>
-              <option value="その他">その他</option>
-            </select>
-            {initialPeriod.reason === 'その他' && (
-              <input
-                type="text"
-                placeholder="詳細を入力してください"
-                className="w-full mt-2 p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-                onChange={(e) => onPeriodChange({ ...initialPeriod, reason: `その他: ${e.target.value}` })}
-              />
-            )}
-          </div>
-          
-          <div className="flex justify-end gap-2 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="px-4 py-2 text-gray-600 border rounded hover:bg-gray-50"
-            >
-              キャンセル
-            </button>
-            <button
-              type="submit"
-              className="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700"
-            >
-              使用不能期間を設定
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-};
-// 車両使用不能期間設定モーダルコンポーネント
-                    <div>
-                      <h3 className="text-xl font-bold text-gray-900">トラック稼働状況</h3>
-                      <p className="text-sm text-gray-600">
-                        登録台数: {trucks.length}台 | 
-                        稼働中: {trucks.filter(t => t.status === 'available').length}台 | 
-                        整備中: {trucks.filter(t => t.status === 'maintenance').length}台
-                      </p>
-                    </div>
-                  </div>
-                </div>
-                
-                {trucks.length === 0 ? (
-                  <div className="p-7 text-center">
-                    <p className="text-gray-500 mb-2">登録済みのトラックがありません</p>
-                    <p className="text-sm text-gray-400">トラック登録・編集タブから新しいトラックを追加してください</p>
-                  </div>
-                ) : (
-                  <div className="p-7">
-                    <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-6">
-                      {trucks.map(truck => {
-                        const nextSchedule = truck.schedules
-                          .filter(s => s.date >= new Date().toISOString().split('T')[0])
-                          .sort((a, b) => a.date.localeCompare(b.date))[0];
-                        
-                        const todaySchedules = truck.schedules
-                          .filter(s => s.date === new Date().toISOString().split('T')[0])
-                          .length;
-
-                        return (
-                          <div key={truck.id} className={`border-2 rounded-lg p-4 transition-all hover:shadow-md ${
-                            truck.status === 'available' ? 'border-green-200 bg-green-50' :
-                            truck.status === 'maintenance' ? 'border-yellow-200 bg-yellow-50' :
-                            'border-red-200 bg-red-50'
-                          }`}>
-                            {/* ヘッダー */}
-                            <div className="flex justify-between items-start mb-3">
-                              <div>
-                                <h4 className="font-bold text-gray-900">{truck.name}</h4>
-                                <p className="text-sm text-gray-600">{truck.plateNumber}</p>
-                              </div>
-                              <div className="text-right">
-                                <div className="flex items-center gap-2 mb-2">
-                                  <button
-                                    onClick={() => {
-                                      setSelectedTruckForUnavailable(truck);
-                                      setShowUnavailablePeriodModal(true);
-                                    }}
-                                    className="px-2 py-1 text-xs bg-red-100 text-red-700 border border-red-300 rounded hover:bg-red-200 transition-colors"
-                                    title="使用不能期間を設定"
-                                  >
-                                    🚫 使用不能
-                                  </button>
-                                  <span className={`px-3 py-1 text-xs font-bold rounded-full ${getStatusConfig('truck', truck.status).color}`}>
-                                    {getStatusConfig('truck', truck.status).icon} {getStatusConfig('truck', truck.status).text}
-                                  </span>
-                                </div>
-                                {todaySchedules > 0 && (
-                                  <p className="text-xs text-gray-600 mt-1">本日: {todaySchedules}件</p>
-                                )}
-                              </div>
-                            </div>
-
-                            {/* 基本情報 */}
-                            <div className="grid grid-cols-2 gap-2 mb-3 text-sm">
-                              <div className="flex items-center gap-1">
-                                <span className="text-gray-500">📦</span>
-                                <span className="font-medium">{truck.capacityKg.toLocaleString()}kg</span>
-                              </div>
-                              <div className="flex items-center gap-1">
-                                <span className="text-gray-500">🚛</span>
-                                <span className="text-gray-700">{truck.truckType}</span>
-                              </div>
-                            </div>
-                            
-                            {/* 次回稼働予定 */}
-                            {nextSchedule ? (
-                              <div className={`p-3 rounded-lg border-l-4 ${
-                                nextSchedule.contractStatus === 'confirmed' 
-                                  ? 'bg-blue-50 border-blue-400' 
-                                  : 'bg-orange-50 border-orange-400'
-                              }`}>
-                                <div className="flex justify-between items-start mb-1">
-                                  <p className="text-sm font-bold text-gray-900">次回稼働予定</p>
-                                  <span className={`px-2 py-0.5 text-xs font-medium rounded ${getStatusConfig('schedule', nextSchedule.contractStatus || 'estimate').color}`}>
-                                    {getStatusConfig('schedule', nextSchedule.contractStatus || 'estimate').icon} {getStatusConfig('schedule', nextSchedule.contractStatus || 'estimate').text}
-                                  </span>
-                                </div>
-                                <p className="text-sm font-medium text-gray-800">
-                                  📅 {new Date(nextSchedule.date).toLocaleDateString('ja-JP', { 
-                                    month: 'short', 
-                                    day: 'numeric',
-                                    weekday: 'short' 
-                                  })}
-                                </p>
-                                <p className="text-sm text-gray-700">
-                                  ⏰ {nextSchedule.startTime}-{nextSchedule.endTime}
-                                </p>
-                                {nextSchedule.customerName && (
-                                  <p className="text-sm text-gray-700 mt-1">
-                                    👤 {nextSchedule.customerName}様
-                                  </p>
-                                )}
-                                {nextSchedule.workType && (
-                                  <p className="text-xs text-gray-600 mt-1">
-                                    作業: {nextSchedule.workType === 'loading' ? '積み込み' :
-                                           nextSchedule.workType === 'unloading' ? '荷下ろし' :
-                                           nextSchedule.workType === 'moving' ? '輸送' : '整備'}
-                                  </p>
-                                )}
-                              </div>
-                            ) : (
-                              <div className="p-3 rounded-lg bg-gray-100 border-l-4 border-gray-300">
-                                <p className="text-sm font-medium text-gray-600">📋 稼働予定なし</p>
-                                <p className="text-xs text-gray-500 mt-1">新しい案件をアサイン可能</p>
-                              </div>
-                            )}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                )}
-              </div>
-
-              {/* 案件一覧 */}
-              <div className="bg-white shadow rounded-lg">
-                <div className="px-4 py-5 sm:p-6">
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center gap-4">
-                      <div className="w-14 h-14 bg-gray-100 rounded-xl flex items-center justify-center text-3xl">📋</div>
-                      <div>
-                        <h3 className="text-xl font-bold text-gray-600">案件詳細</h3>
-                        <p className="text-sm text-gray-900">引越し案件の管理・編集</p>
-                      </div>
-                    </div>
-                    {selectedCaseId && registrationMode === 'registration' && (
-                      <div className="bg-green-100 border border-green-300 rounded-lg p-3">
-                        <div className="flex items-center gap-2">
-                          <span className="text-green-600 text-lg">🎯</span>
-                          <span className="text-sm font-medium text-green-800">
-                            新規登録案件が緑色でハイライト表示されています
-                          </span>
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                </div>
-                
-                {formSubmissions.length === 0 ? (
-                  <div className="p-7 text-center">
-                    <p className="text-gray-500 mb-4">入力フォームから送信された案件がありません</p>
-                    <p className="text-sm text-gray-400">
-                      顧客が引越し見積もりフォームを送信すると、ここに表示されます
-                    </p>
-                  </div>
-                ) : (
-                  <div className="p-7 space-y-6">
-                    {formSubmissions.map(submission => {
-                      const isExpanded = expandedSubmissions.has(submission.id);
-                      
-                      const toggleExpanded = () => {
-                        const newExpandedSubmissions = new Set(expandedSubmissions);
-                        if (isExpanded) {
-                          newExpandedSubmissions.delete(submission.id);
-                        } else {
-                          newExpandedSubmissions.add(submission.id);
-                        }
-                        setExpandedSubmissions(newExpandedSubmissions);
-                      };
-                      
-                      // 新規登録された案件かどうかを判定
-                      const isNewlyRegistered = submission.id === selectedCaseId && registrationMode === 'registration';
-                      
-                      return (
-                        <div 
-                          key={submission.id} 
-                          className={`bg-white rounded-xl shadow border-2 hover:shadow-lg transition-all duration-300 ${
-                            isNewlyRegistered 
-                              ? 'border-green-400 bg-green-50 ring-2 ring-green-200 animate-pulse' 
-                              : 'border-gray-200'
-                          }`}
-                        >
-                        <div className="p-6">
-                          <div className="flex justify-between items-start mb-4">
-                            <div className="flex items-center gap-4">
-                              <div>
-                                <div className="flex items-center gap-2">
-                                  <h3 className="text-lg font-bold text-gray-900">{submission.customerName}</h3>
-                                  {isNewlyRegistered && (
-                                    <span className="px-2 py-1 text-xs font-bold bg-green-500 text-white rounded-full animate-bounce">
-                                      🆕 新規登録
-                                    </span>
-                                  )}
-                                </div>
-                                <div className="flex items-center gap-4 text-sm">
-                                  <p className="text-gray-600">{formatDate(submission.moveDate)}</p>
-                                  <div className="flex items-center gap-2">
-                                    <span className="text-blue-600 font-medium">
-                                      📋 {submission.requestSource || (submission.isManualRegistration ? '手動登録' : 'シンクワーク')}
-                                    </span>
-                                    {submission.isManualRegistration && submission.customerPhone && (
-                                      <span className="text-gray-600">
-                                        📞 {submission.customerPhone}
-                                      </span>
-                                    )}
-                                  </div>
-                                </div>
-                              </div>
-                              <div className="flex items-center gap-2 text-sm">
-                                <span className="font-semibold text-blue-600 bg-blue-50 px-3 py-1 rounded-full">
-                                  📊 {submission.totalPoints}pt
-                                </span>
-                                <span className="font-semibold text-green-600 bg-green-50 px-3 py-1 rounded-full">
-                                  💰 ¥{calculateEstimatedPrice(submission.totalPoints, submission.distance || 0).toLocaleString()}
-                                </span>
-                              </div>
-                            </div>
-                            <div className="flex items-center gap-2">
-                              <span className={`px-3 py-1 text-xs font-medium rounded-full ${getStatusConfig('submission', submission.status).color}`}>
-                                {getStatusConfig('submission', submission.status).icon} {getStatusConfig('submission', submission.status).text}
-                              </span>
-                              
-                              {/* 案件ステータス */}
-                              {submission.caseStatus && (
-                                <span className={`px-2 py-1 text-xs font-medium rounded-full ${getStatusConfig('caseStatus', submission.caseStatus).color}`}>
-                                  {getStatusConfig('caseStatus', submission.caseStatus).icon} {getStatusConfig('caseStatus', submission.caseStatus).text}
-                                </span>
-                              )}
-                              
-                              {submission.contractStatus === 'estimate' && (
-                                <span className={`px-2 py-1 text-xs font-medium rounded-full ${getStatusConfig('contract', 'estimate').color}`}>
-                                  {getStatusConfig('contract', 'estimate').icon} {getStatusConfig('contract', 'estimate').text}
-                                </span>
-                              )}
-                              {submission.contractStatus === 'confirmed' && (
-                                <span className={`px-2 py-1 text-xs font-medium rounded-full ${getStatusConfig('contract', 'confirmed').color}`}>
-                                  {getStatusConfig('contract', 'confirmed').icon} {getStatusConfig('contract', 'confirmed').text}
-                                </span>
-                              )}
-                              
-                              {/* キャンセルボタン（受注案件のみ表示） */}
-                              {(submission.caseStatus === 'contracted' || submission.contractStatus === 'confirmed') && 
-                               submission.caseStatus !== 'cancelled' && (
-                                <button
-                                  onClick={() => changeCaseStatus(submission.id, 'cancelled')}
-                                  className="px-2 py-1 text-xs bg-yellow-100 text-yellow-800 border border-yellow-300 rounded hover:bg-yellow-200 transition-colors"
-                                  title="受注案件をキャンセルに変更（請求対象）"
-                                >
-                                  🚫 キャンセル
-                                </button>
-                              )}
-                              
-                              <button
-                                onClick={toggleExpanded}
-                                className="text-gray-500 hover:text-gray-700 p-1 rounded"
-                              >
-                                {isExpanded ? '▲ 詳細を閉じる' : '▼ 詳細を表示'}
-                              </button>
-                            </div>
-                          </div>
-
-                          {/* 基本情報の簡潔表示 */}
-                          <div className="grid grid-cols-2 md:grid-cols-3 gap-3 mb-4 text-sm">
-                            <div className="flex items-center gap-2">
-                              <span className="text-gray-500">📦</span>
-                              <span className="font-medium">{submission.totalCapacity.toLocaleString()}kg</span>
-                            </div>
-                            <div className="flex items-center gap-2">
-                              <span className="text-gray-500">📍</span>
-                              <span className="text-gray-700 truncate">{submission.originAddress}</span>
-                            </div>
-                            <div className="flex items-center gap-2">
-                              <span className="text-gray-600">🏁</span>
-                              <span className="text-gray-700 truncate">{submission.destinationAddress}</span>
-                            </div>
-                          </div>
-
-                          {/* 推奨トラックと割り当て状況（簡潔版） */}
-                          <div className="flex items-center justify-between">
-                            <div className="flex items-center gap-2">
-                              <span className="text-sm text-gray-600">推奨:</span>
-                              <div className="flex gap-1">
-                                {calculateRecommendedTrucks(submission.totalPoints).slice(0, 3).map(truck => (
-                                  <span key={truck.id} className="text-xs bg-blue-50 text-blue-700 px-2 py-1 rounded">
-                                    {truck.truckType}
-                                  </span>
-                                ))}
-                              </div>
-                            </div>
-                            <div className="flex items-center gap-2">
-                              {submission.truckAssignments.length > 0 && (
-                                <span className="text-xs text-green-600 font-medium">
-                                  🚚 {submission.truckAssignments.length}台割当済
-                                </span>
-                              )}
-                              {submission.status !== 'completed' && (
-                                <button
-                                  onClick={() => {
-                                    setSelectedSubmission(submission);
-                                    setShowTruckModal(true);
-                                  }}
-                                  className="bg-blue-600 hover:bg-blue-700 text-white px-3 py-1 rounded text-xs font-medium"
-                                >
-                                  + トラック割当
-                                </button>
-                              )}
-                            </div>
-                          </div>
-                        </div>
-
-                        {/* 展開可能な詳細情報 */}
-                        {isExpanded && (
-                          <div className="border-t bg-gray-50 p-6 space-y-4">
-                            {/* 連絡先情報 */}
-                            <div>
-                              <h4 className="text-sm font-semibold text-gray-900 mb-2">連絡先情報</h4>
-                              <div className="grid grid-cols-2 gap-3 text-sm">
-                                <div>📧 {submission.customerEmail}</div>
-                                <div>📞 {submission.customerPhone}</div>
-                              </div>
-                            </div>
-                            {/* 荷物詳細 */}
-                            <div>
-                              <h4 className="text-sm font-semibold text-gray-900 mb-2">荷物リスト</h4>
-                              <div className="flex flex-wrap gap-1">
-                                {submission.itemList.map((item, index) => (
-                                  <span key={index} className="text-xs bg-white border px-2 py-1 rounded">
-                                    {item}
-                                  </span>
-                                ))}
-                              </div>
-                            </div>
-
-                            {/* 追加サービス */}
-                            {submission.additionalServices.length > 0 && (
-                              <div>
-                                <h4 className="text-sm font-semibold text-gray-900 mb-2">追加サービス</h4>
-                                <div className="flex flex-wrap gap-1">
-                                  {submission.additionalServices.map((service, index) => (
-                                    <span key={index} className="text-xs bg-blue-100 text-blue-800 px-2 py-1 rounded">
-                                      {service}
-                                    </span>
-                                  ))}
-                                </div>
-                              </div>
-                            )}
-
-                            {/* 割り当てトラック */}
-                            {submission.truckAssignments.length > 0 && (
-                              <div>
-                                <h4 className="text-sm font-semibold text-gray-900 mb-2">割り当てトラック</h4>
-                                <div className="space-y-2">
-                                  {submission.truckAssignments.map((truckAssignment, index) => (
-                                    <div key={index} className="flex justify-between items-center p-2 bg-white rounded border">
-                                      <div>
-                                        <span className="font-medium text-sm">{truckAssignment.truckName}</span>
-                                        <span className="text-xs text-gray-600 ml-2">
-                                          {formatTime(truckAssignment.startTime)}-{formatTime(truckAssignment.endTime)} 
-                                          ({truckAssignment.capacity.toLocaleString()}kg)
-                                        </span>
-                                      </div>
-                                      {submission.status !== 'completed' && (
-                                        <button
-                                          onClick={() => removeTruckFromSubmission(submission.id, truckAssignment.truckId)}
-                                          className="text-red-600 hover:text-red-800 text-xs px-2 py-1 hover:bg-red-50 rounded"
-                                        >
-                                          削除
-                                        </button>
-                                      )}
-                                    </div>
-                                  ))}
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        )}
-                      </div>
-                    );
-                  })}
-                  </div>
-                )}
-              </div>
-            </>
-          )}
-          
-          {activeTab === 'registration' && (
-            <div className="bg-white shadow rounded-lg">
-              <div className="px-4 py-5 sm:p-6">
-                <div className="flex items-center gap-4">
-                  <div className="w-14 h-14 bg-purple-100 rounded-xl flex items-center justify-center text-3xl">🚚</div>
-                  <div>
-                    <h3 className="text-xl font-bold text-purple-600">トラック登録・編集</h3>
-                    <p className="text-sm text-gray-900">車両情報の管理・更新</p>
-                  </div>
-                </div>
-              </div>
-              <div className="px-4 py-5 sm:p-6">
-                <TruckRegistration
-                  trucks={trucks}
-                  selectedTruck={selectedTruck}
-                  onAddTruck={addTruck}
-                  onUpdateTruck={updateTruck}
-                  onDeleteTruck={deleteTruck}
-                  onSelectTruck={setSelectedTruck}
-                  availableTruckTypes={availableTruckTypes}
-                  pricingRules={pricingRules}
-                />
-              </div>
-            </div>
-          )}
-        </div>
-      </main>
-
-      {/* モーダル */}
-      {showTruckModal && (
-        <TruckAssignmentModal
-          selectedSubmission={selectedSubmission}
-          trucks={trucks}
-          pricingTrucks={pricingTrucks}
-          setShowTruckModal={setShowTruckModal}
-          assignTruckToSubmission={assignTruckToSubmission}
-          calculateRecommendedTrucks={calculateRecommendedTrucks}
-          calculateEstimatedPrice={calculateEstimatedPrice}
-        />
-      )}
-
-      {/* 車両使用不能期間設定モーダル */}
-      {showUnavailablePeriodModal && (
-        <UnavailablePeriodModal
-          truck={selectedTruckForUnavailable}
-          onClose={() => {
-            setShowUnavailablePeriodModal(false);
-            setSelectedTruckForUnavailable(null);
-            setUnavailablePeriod({ startDate: '', endDate: '', reason: '' });
-          }}
-          onSave={(period) => {
-            if (!selectedTruckForUnavailable) return;
-            
-            // 指定期間に使用不能スケジュールを作成
-            const startDate = new Date(period.startDate);
-            const endDate = new Date(period.endDate);
-            const newSchedules: Schedule[] = [];
-            
-            for (let date = new Date(startDate); date <= endDate; date.setDate(date.getDate() + 1)) {
-              newSchedules.push({
-                id: `unavailable-${crypto.randomUUID()}`,
-                date: toLocalDateString(date),
-                startTime: '00:00',
-                endTime: '23:59',
-                status: 'maintenance',
-                customerName: '',
-                workType: 'maintenance',
-                description: `使用不能期間: ${period.reason}`,
-                capacity: 0,
-                origin: '',
-                destination: '',
-              });
-            }
-            
-            const updatedTruck = {
-              ...selectedTruckForUnavailable,
-              schedules: [...selectedTruckForUnavailable.schedules, ...newSchedules]
-            };
-            
-            updateTruck(updatedTruck);
-            setShowUnavailablePeriodModal(false);
-            setSelectedTruckForUnavailable(null);
-            setUnavailablePeriod({ startDate: '', endDate: '', reason: '' });
-          }}
-          initialPeriod={unavailablePeriod}
-          onPeriodChange={setUnavailablePeriod}
-        />
-      )}
-    </div>
-  );
-}
-
-// 車両使用不能期間設定モーダルコンポーネント
-interface UnavailablePeriodModalProps {
-  truck: Truck | null;
-  onClose: () => void;
-  onSave: (period: { startDate: string; endDate: string; reason: string }) => void;
-  initialPeriod: { startDate: string; endDate: string; reason: string };
-  onPeriodChange: (period: { startDate: string; endDate: string; reason: string }) => void;
-}
-
-const UnavailablePeriodModal = ({ truck, onClose, onSave, initialPeriod, onPeriodChange }: UnavailablePeriodModalProps) => {
-  if (!truck) return null;
-  
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    if (!initialPeriod.startDate || !initialPeriod.endDate) {
-      alert('開始日と終了日を入力してください');
-      return;
-    }
-    
-    if (new Date(initialPeriod.startDate) > new Date(initialPeriod.endDate)) {
-      alert('開始日は終了日より前の日付を選択してください');
-      return;
-    }
-    
-    onSave(initialPeriod);
-  };
-  
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md">
-        <h3 className="text-lg font-semibold mb-4">車両使用不能期間設定</h3>
-        <p className="text-sm text-gray-600 mb-4">
-          対象車両: {truck.name} ({truck.plateNumber})
-        </p>
-        
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              開始日 <span className="text-red-500">*</span>
-            </label>
-            <input
-              type="date"
-              value={initialPeriod.startDate}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, startDate: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-              required
-            />
-          </div>
-          
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              終了日 <span className="text-red-500">*</span>
-            </label>
-            <input
-              type="date"
-              value={initialPeriod.endDate}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, endDate: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-              required
-            />
-          </div>
-          
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">
-              理由
-            </label>
-            <select
-              value={initialPeriod.reason}
-              onChange={(e) => onPeriodChange({ ...initialPeriod, reason: e.target.value })}
-              className="w-full p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              <option value="">選択してください</option>
-              <option value="車検">車検</option>
-              <option value="定期点検">定期点検</option>
-              <option value="修理">修理</option>
-              <option value="清掃・メンテナンス">清掃・メンテナンス</option>
-              <option value="休車">休車</option>
-              <option value="その他">その他</option>
-            </select>
-            {initialPeriod.reason === 'その他' && (
-              <input
-                type="text"
-                placeholder="詳細を入力してください"
-                className="w-full mt-2 p-2 border rounded focus:outline-none focus:ring-2 focus:ring-blue-500"
-                onChange={(e) => onPeriodChange({ ...initialPeriod, reason: `その他: ${e.target.value}` })}
-              />
-            )}
-          </div>
-          
-          <div className="flex justify-end gap-2 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="px-4 py-2 text-gray-600 border rounded hover:bg-gray-50"
-            >
-              キャンセル
-            </button>
-            <button
-              type="submit"
-              className="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700"
-            >
-              使用不能期間を設定
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-};
 
 export default function DispatchManagement() {
   return (
